Started working on the automove (autonomous function).
Worked out most of the logic of the function. Now, I need to write in the rampdown part of the function.

Soon, I will finish up most of the logic. Once I have gone and tested it on the 10B Robot, I will push these commits to the master branch.

It looks promising. I drew some diagrams, thought some of it out!

I'm so excited!

Its very well commented and the uppercase comments are the spots where more code is required. Also, i need to come up with a defining boolean for the while loop.